+++ actix-limitation/CHANGES.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-- Update `redis` dependency to `0.26`.
-- Update `actix-session` dependency to `0.9`.
+- Update `redis` dependency to `1`.
+- Update `actix-session` dependency to `0.11`.
+- Minimum supported Rust version (MSRV) is now 1.88.
+- Add `redis` TLS features to `actix-limitation`.
 
 ## 0.5.1