+++ actix-protobuf/CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Updated `prost` dependency to `0.14`.
+- Minimum supported Rust version (MSRV) is now 1.88.
+
 ## 0.11.0
 
 - Updated `prost` dependency to `0.13`.